Fix environment sampling bug
The environment color is now accumulated last instead of first (which was the bug).
Previously the environment was sampled based on the ray start rather than using
the final bounced ray's orientation.

With the current env gradient, the bug is hard to notice, but appears in the dielectric
and metal reflections.