Accumulate color in rays
Color is now accumulated directly in each ray, rather than in the final texture.
This both resolves a race condition where multiple threads could update the texture
at the same time and resolves a bug in older hardware which was not capable of
accessing float4 via RWTexture2D, which was introduced in DirectX 11.3

Also fixed bug where _MaxBounces was set as a float instead of an int, which was
causing MaxBounces to not actuall limit the number of bounces per ray.